Move Profile to its final spot
